all addressing modes implemented in assembler

+++ assembler.h
@@ -11,7 +11,7 @@ public:
   Assembler(Memory& memory, uint16_t origin = 0);
   void setOrigin(uint16_t origin) { address_ = origin; }
   auto currentAddress() const { return address_; }
-  bool assemble(InstructionType type, AddressingMode mode = NoOperands, uint16_t operand = 0);
+  bool assemble(InstructionType type, AddressingMode mode = NoOperands, int operand = 0);
   bool assemble(const char*);
 
 private:
